pyproject.toml
[project]
name = "testorderservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aio-pika>=9.5.5",
    "alembic>=1.15.1",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "celery>=5.4.0",
    "email-validator>=2.2.0",
    "fastapi-limiter>=0.1.6",
    "fastapi[standard]>=0.115.11",
    "fntypes>=0.1.4.post3",
    "greenlet>=3.1.1",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "redis[hiredis]>=5.2.1",
    "sqlalchemy>=2.0.39",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.11.0",
]
